% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_shiny.R
\name{loess_shiny_server}
\alias{loess_shiny_server}
\title{Shiny UI}
\usage{
loess_shiny_server(input, output, session)
}
\arguments{
\item{input}{shiny input object.}

\item{output}{shiny output object.}

\item{session}{shiny session object.}
}
\value{
a Shiny UI object.
}
\description{
Shiny UI
}
